Extract shared methods
