+++ syft/serde/torch/serde.py
@@ -1,4 +1,11 @@
+import io
+from tempfile import TemporaryFile
+import warnings
+
+import numpy
 import torch
+
+from syft.workers.abstract import AbstractWorker
 
 # Torch dtypes to string (and back) mappers
 TORCH_DTYPE_STR = {
@@ -20,3 +27,44 @@ TORCH_DTYPE_STR = {
     torch.bfloat16: "bfloat16",
 }
 TORCH_STR_DTYPE = {name: cls for cls, name in TORCH_DTYPE_STR.items()}
+
+
+def torch_tensor_serializer(worker: AbstractWorker, tensor) -> bin:
+    """Strategy to serialize a tensor using Torch saver"""
+    binary_stream = io.BytesIO()
+    torch.save(tensor, binary_stream)
+    return binary_stream.getvalue()
+
+
+def torch_tensor_deserializer(worker: AbstractWorker, tensor_bin) -> torch.Tensor:
+    """"Strategy to deserialize a binary input using Torch load"""
+    bin_tensor_stream = io.BytesIO(tensor_bin)
+    return torch.load(bin_tensor_stream)
+
+
+def numpy_tensor_deserializer(worker: AbstractWorker, tensor_bin) -> torch.Tensor:
+    """"Strategy to deserialize a binary input in npy format into a Torch tensor"""
+    input_file = TemporaryFile()
+    input_file.write(tensor_bin)
+    # read data from file
+    input_file.seek(0)
+    return torch.from_numpy(numpy.load(input_file))
+
+
+def numpy_tensor_serializer(worker: AbstractWorker, tensor: torch.Tensor) -> bin:
+    """Strategy to serialize a tensor using numpy npy format.
+    If tensor requires to calculate gradients, it will be detached.
+    """
+    if tensor.requires_grad:
+        warnings.warn(
+            "Torch to Numpy serializer can only be used with tensors that do not require grad. "
+            "Detaching tensor to continue"
+        )
+        tensor = tensor.detach()
+
+    np_tensor = tensor.numpy()
+    outfile = TemporaryFile()
+    numpy.save(outfile, np_tensor)
+    # Simulate close and open by calling seek
+    outfile.seek(0)
+    return outfile.read()
